Big Refactory:
For more flexiable column store:
1. Remove ReadonlySegment::m_part / m_nonIndexSchema ....
2. Uniform colgroup and Index+Store, Index+Store are front
   indexNum colgroups, the Store-only Colgroup are following
3. Refactory dispatch of store file types
4. All these changes are prepair for real column store

+++ src/nark/db/db_context.hpp
@@ -35,7 +35,7 @@ public:
 	valvec<byte> row2;
 	valvec<byte> key1;
 	valvec<byte> key2;
-	valvec<size_t> offsets;
+	valvec<uint32_t> offsets;
 	valvec<fstring> cols1;
 	valvec<fstring> cols2;
 	bool syncIndex;
